Formatter and interface update
WMS has  a new formatter, capable of filtering more layers (top level and one extra level deeper).
Amsterdam WMS & WFS data now also supported (due to less shady way of reading layers)
Separated the WMS and WFS reading features into their own interfaces and make assumptions about the link based on this.

+++ Packages/Netherlands3D/WMSLayer/Runtime/Scripts/BoundingBox.cs
@@ -5,14 +5,14 @@ using UnityEngine;
 [System.Serializable]
 public struct BoundingBox
 {
-    public float MinX;
-    public float MinY;
-    public float MaxX;
-    public float MaxY;
+    public double MinX;
+    public double MinY;
+    public double MaxX;
+    public double MaxY;
 
     public static readonly BoundingBox Zero = new BoundingBox(0, 0, 0, 0);
 
-    public BoundingBox(float minX, float minY, float maxX, float maxY)
+    public BoundingBox(double minX, double minY, double maxX, double maxY)
     {
         MinX = minX;
         MinY = minY;
